Move core duplicate processing to shared lib
